Generate proper sensor outputs
Due to integer / double issues, I was either getting 0 or 1. Now I've changed
the members of UltrasonicSensor that were causing the problem to doubles.

+++ src/ultrasonic_sensor.hpp
@@ -16,11 +16,11 @@
  * input device.
  */
 class UltrasonicSensor {
-	int min_range;
-	int max_range;
+	double min_range;
+	double max_range;
 	int serial_connection;
 public:
-	UltrasonicSensor(const int &min, const int &max, const int &serial);
+	UltrasonicSensor(const double &min, const double &max, const int &serial);
 	double detect();
 };
